+++ examples/mocha/extension.js
@@ -1,0 +1,16 @@
+const { Criteria } = require("kelonio");
+
+module.exports = {
+    extension: {
+        extraReport: benchmark => {
+            const fastest = benchmark.find(Criteria.Fastest);
+            if (fastest) {
+                return `
+= = Custom Report = =
+Fastest: "${fastest.description.join("/")}" (${fastest.mean} ms)
+= = = = = = = = = = =
+        `;
+            }
+        }
+    }
+};