+++ tests/extension.js
@@ -1,0 +1,12 @@
+const { Criteria } = require("kelonio");
+
+module.exports = {
+    extension: {
+        extraReport: benchmark => {
+            const fastest = benchmark.find(Criteria.Fastest);
+            if (fastest) {
+                return `Fastest: "${fastest.description.join("/")}" (${fastest.mean} ms)`;
+            }
+        },
+    },
+};